pyproject.toml
[tool.poetry]
authors = ["Leandro Lucarella <luca-frequenz@llucax.com>"]
name = "python-grpc-benchmark"
version = "0.1.0"
description = ""
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
protobuf = "~5.26.1"             # not using ^ to avoid libprotobuf 27.x
grpcio = "~1.64.1"
grpcio-tools = "~1.64.1"
grpclib = "^0.4.7"
click = "^8.1.7"
psutil = "^6.0.0"
types-psutil = "^6.0.0.20240621"
matplotlib = "^3.9.0"
pyqt6 = "^6.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
